Move Split method from BaseJellyfinApiController.cs to RequestHelpers.cs
